refactor(ui): change folder structure
