+++ packages/ui/src/index.ts
@@ -1,1 +1,1 @@
-export * from "./Button"
+export * from "./components/Button/Button"
